Simplify main class

+++ src/main/java/me/bechberger/Main.java
@@ -7,6 +7,8 @@ import java.io.*;
 import java.lang.instrument.Instrumentation;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.StandardCopyOption;
+import java.nio.file.StandardOpenOption;
 import java.util.jar.JarFile;
 
 /**
@@ -40,31 +42,15 @@ public class Main {
         System.exit(new CommandLine(new Processor()).execute(args));
     }
 
-    private static Path getExtractedJARPath() {
-        try {
-            // based on https://github.com/gkubisa/jni-maven/blob/master/src/main/java/ie/agisoft/LibraryLoader.java
-            InputStream in = Main.class.getClassLoader().getResourceAsStream("dead-code-runtime.jar");
-            assert in != null;
-
-            File file = File.createTempFile("runtime", ".jar");
-
-            file.deleteOnExit();
-            try {
-                byte[] buf = new byte[4096];
-                try (OutputStream out = new FileOutputStream(file)) {
-                    while (in.available() > 0) {
-                        int len = in.read(buf);
-                        if (len >= 0) {
-                            out.write(buf, 0, len);
-                        }
-                    }
-                }
-            } finally {
-                in.close();
+    private static Path getExtractedJARPath() throws IOException {
+        try (InputStream in = Main.class.getClassLoader().getResourceAsStream("dead-code-runtime.jar")){
+            if (in == null) {
+                throw new RuntimeException("Could not find dead-code-runtime.jar");
             }
+            File file = File.createTempFile("runtime", ".jar");
+            file.deleteOnExit();
+            Files.copy(in, file.toPath(), StandardCopyOption.REPLACE_EXISTING);
             return file.toPath().toAbsolutePath();
-        } catch (IOException e) {
-            throw new RuntimeException(e);
         }
     }
 }
